requirements.txt
anthropic==0.80.0
httpx==0.28.1
feedparser==6.0.12
python-dotenv==1.2.1
beautifulsoup4==4.14.3
